IP: Moving and renaming files to match style guide.
