+++ image_proc/CMakeLists.txt
@@ -155,4 +155,9 @@ ament_export_libraries(
   crop_decimate
 )
 
+if(BUILD_TESTING)
+  find_package(ament_lint_auto REQUIRED)
+  ament_lint_auto_find_test_dependencies()
+endif()
+
 ament_package()
