+++ image_proc/include/image_proc/debayer.hpp
