+++ image_proc/include/image_proc/edge_aware.hpp
