+++ image_proc/include/image_proc/resize.hpp
